+++ src/test/java/com/commercetools/sync/producttypes/helpers/ProductTypeReferenceResolverTest.java
@@ -30,7 +30,7 @@ import static org.mockito.Mockito.when;
 class ProductTypeReferenceResolverTest {
 
     @Test
-    void resolveReferences_WithNullAttributes_ShouldResolveReferences() {
+    void resolveReferences_WithNullAttributes_ShouldNotResolveReferences() {
         // preparation
         final ProductTypeDraft productTypeDraft =
             ProductTypeDraftBuilder.of("foo", "foo", "desc", null)
@@ -87,7 +87,6 @@ class ProductTypeReferenceResolverTest {
     @Test
     void resolveReferences_WithNoNestedTypeReferences_ShouldResolveReferences() {
         // preparation
-
         final AttributeDefinitionDraft attributeDefinitionDraft = AttributeDefinitionDraftBuilder
             .of(StringAttributeType.of(), "string attr", ofEnglish("string attr label"), true)
             .build();